requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas==2.1.3
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
aiofiles==23.2.1
redis==5.0.1
rq==1.15.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
